Advance the org.mobilitydb.meos facade to the MobilityDB master MEOS surface
Apply the 77-symbol drift from pin 588768d7ab to master: 20 trgeo→trgeometry
renames (A); 12 pg_*/cstring2text/text2cstring/pose_normalise/temporal_instants_p/
temporal_hausdorff_distance renames (B); 21 removed scalar helpers dropped with
no callers (C: add_date_int, add_interval_interval, add_timestamptz_interval,
date_to_timestamp, date_to_timestamptz, float_exp, float_ln, float_log10,
float_round, int32_cmp, int64_cmp, interval_make, mul_interval_double,
timestamp_to_date, timestamptz_shift, timestamptz_to_date, minus_date_date,
minus_date_int, minus_timestamptz_interval, minus_timestamptz_timestamptz,
meos_initialize_pointcloud); 23 signature updates (D): value_at_timestamptz
gains a result out-param and returns bool; value_split/value_time_split/
temporal_time_split gain bins/count out-params; set_spans/spanset_spans/
spanset_spanarr drop the count out-param; super_union_span_span drops the
boolean merge flag; text_cmp gains a collation pointer; gbox_make gains hasm
and geodetic booleans and min/max range pairs; geom_min_bounding_radius gains
a radius out-param; pose_make_2d/3d gain a geodetic boolean; trgeoinst_make is
renamed trgeometryinst_make. All xxx_hash_extended functions change their seed
and return type from long to int (bigint_to_span/spanset likewise). 1735 tests
pass, BUILD SUCCESS.

+++ jmeos-core/src/main/java/org/mobilitydb/meos/MeosOpsFreeNpoint.java
@@ -177,7 +177,7 @@ public final class MeosOpsFreeNpoint {
      * <p>Pure per-event; safe in any scalar position.</p>
      * <p>Free function (not class-classified). Classification rule: scalar comparison/hash</p>
      */
-    public static long npoint_hash_extended(jnr.ffi.Pointer arg0, long arg1) {
+    public static int npoint_hash_extended(jnr.ffi.Pointer arg0, int arg1) {
         if (!MEOS_AVAILABLE) {
             throw new UnsupportedOperationException(
                 "npoint_hash_extended requires libmeos — set -Dmeos.enabled=true");
