+++ jmeos-core/src/main/java/org/mobilitydb/meos/MeosOpsSpan.java
@@ -158,7 +158,7 @@ public final class MeosOpsSpan {
      * <p>Object-model role: {@code accessor}.</p>
      * <p>Classification: role=accessor</p>
      */
-    public static long span_hash_extended(jnr.ffi.Pointer arg0, long arg1) {
+    public static int span_hash_extended(jnr.ffi.Pointer arg0, int arg1) {
         if (!MEOS_AVAILABLE) {
             throw new UnsupportedOperationException(
                 "span_hash_extended requires libmeos — set -Dmeos.enabled=true");
